Network player animations and fix collisions
