+++ Assets/Scripts/Networking/NetworkedObjectTransform.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c0726d719c69cec4ca7a806c6d7effbd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
